backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
backoff==2.2.1
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.122.0
google==3.0.0
google-auth==2.43.0
google-genai==1.52.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
jiter==0.12.0
langfuse==3.10.1
openai==2.8.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
postgrest==0.18.0
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
starlette==0.50.0
supabase==2.10.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0